+++ splunk_add_on_ucc_framework/commands/build.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 #
 import configparser
+import glob
 import json
 import logging
 import os
@@ -195,30 +196,32 @@ def _get_ignore_list(
                 os.path.join(output_directory, addon_name, utils.get_os_path(path))
             ).strip()
             for path in ignore_list
+            if path.strip()
         ]
         return ignore_list
 
 
-def _remove_listed_files(ignore_list: List[str]) -> None:
+def _remove_listed_files(ignore_list: List[str]) -> List[str]:
     """
     Return path of files/folders to removed in output folder.
 
     Args:
-        ignore_list (list): List of files/folder to removed in output directory.
-
+        ignore_list (list): List of files/folder patterns to be removed in output directory.
     """
-    for path in ignore_list:
-        if os.path.exists(path):
-            if os.path.isfile(path):
-                os.remove(path)
-            elif os.path.isdir(path):
-                shutil.rmtree(path, ignore_errors=True)
-        else:
-            logger.warning(
-                "While ignoring the files mentioned in .uccignore {} was not found".format(
-                    path
-                )
-            )
+    removed_list = []
+    for pattern in ignore_list:
+        paths = glob.glob(pattern, recursive=True)
+        if not paths:
+            logger.warning(f"No files found for the specified pattern: {pattern}")
+            continue
+        for path in paths:
+            if os.path.exists(path):
+                if os.path.isfile(path):
+                    os.remove(path)
+                elif os.path.isdir(path):
+                    shutil.rmtree(path, ignore_errors=True)
+                removed_list.append(path)
+    return removed_list
 
 
 def generate_data_ui(
@@ -629,9 +632,9 @@ def generate(
         os.path.abspath(os.path.join(source, os.pardir, ".uccignore")),
         output_directory,
     )
-    _remove_listed_files(ignore_list)
-    if ignore_list:
-        logger.info(f"Removed {ignore_list} files")
+    removed_list = _remove_listed_files(ignore_list)
+    if removed_list:
+        logger.info("Removed:\n{}".format("\n".join(removed_list)))
     utils.recursive_overwrite(source, os.path.join(output_directory, ta_name))
     logger.info("Copied package directory")